+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ## [Unreleased]
 ### Added
+- custom popup templates
 - raster layers
 - layers can be filtered by match of keywords
 - layers can be filtered by multiple filters